+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Motor enable and disable features in the virtual drive.
 - Emulate control loops in the virtual drive.
 - Support to Python 3.9 to 3.12.
+- EtherCAT PDO module.
 
 ### Deprecated
 - Support to Python 3.6 to 3.8.